Added game
This is a simple game built using HTML, CSS, and JavaScript. The player must click on randomly generated boxes on the game board to score points. Each time a box is clicked, it disappears, and a new one appears in a random location on the game board. The player has 10 seconds to score as many points as possible. The best score is saved locally in the browser's storage.